backend/requirements.txt
Flask==2.3.3
Flask-MongoEngine==1.0.0
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
pymongo==4.6.0
mongoengine==0.27.0
bcrypt==4.1.2

